Project/django-docker/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
django-extensions
werkzeug
pyOpenSSL
PyJWT==2.1.0
Bandit==1.7.0
Pillow
faker
djangorestframework-simplejwt==5.1.0
